Lift parse_prerequisites into shared condition_parser helper
Extracts the structure_api::parse_prerequisites function into
condition_parser::parse_prerequisites_field so it can be reused by
building, hull, and module parsers. No behavior change — structure_api
now calls the shared helper with an identical implementation.

Refs #226
